web/requirements.txt
keras==2.11.0
numpy==1.23.5
Pillow==9.3.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
